A beginning is a very delicate time...
The main players in this are corpus-passed/*, t/01-basic.t and
lib/Language/Picat/Grammar.pm6. The grammar is currently set up to
blindly recognize ("parse" is too strong of a word) the two sample
files.

It just looks for the entire string, makes sure we've reached the end
of the file in both cases, and then reports a successful match. The
'$' term may seem extraneous, but it's very important indeed. It
probably should be factored out, but that's too much trouble at this
point.

I've also added 'use Grammar::Tracer' at the start, along with a
workaround of 'no precompilation' to take care of a tiny issue on my
Linux box that actually bugged me for about a week.

In any case, we can be sure that the grammar in question "parses" the
entire set of files.