Quick fix at last...
Okay, back to what I was explaining.

I've broken up the file *very* crudely at what I think "logical boundaries"
are going to be in the picat language.

I see that ',' is used in place of semicolons, and '.' looks like it
terminates a block. '%...' looks like it's a comment, as is '/*' .. '*/'.
The familiar '=>' is used to separate name from body, and %-style comments
can fly from the end.

All good things to know but I'm most concerned right now where text
starts and stops. So I'll go through the simple boundaries, not
forgetting the comments, and make sure things work.

You'll note as well that I've added this around the text blocks:

^ [ ... || ... ] $

That way I can just add another '||' and the text in case I need to
work with a different file. The ^ .. $ makes sure the parser goes
through the entire text file.