Add the other style of comment.
Now, you'll notice a few things here. I've replaced every comment,
flying or otherwise, with the <comment> block. And where the original
file had '% foo % bar', because I'm going to assume that it's C++-style
and is valid until the end of the line, I'll turn that into a single
<comment> tag.

This is quite a bit of work, and I'd definitely suggest learning either
rectangular edit mode in vim, or doing a careful search-and-replace on
those terms. For the moment I'm going to just do the replacement and
not collapse the <comment>s into <comment>+.