Aand the first real fix.
There are probably other ways to fix this problem, but I've
chosen this simply because it's expedient, and while it could
bite me later I don't think it'll be a problem.

It also gives me a chance to explain the crucial (if not only)
difference between tokens and rules. Suppose for a moment we had
'main' '=>' 'go' as part of our rule. We don't yet, but bear with
me.

Inside the rule, this matches 'main => go', 'main     => go' or
'main=>go', with varying amounts of whitespace between the terms.
Rules effectively insert \s* between each term for us and make things
more convenient, so we don't have to add that manually and account
for each kind of whitespace.

Tokens don't. They're meant for the "lower level" things like
'12.375', '$name' and even '"foo bar"' Each of these forms a single
unit, in fact you could call them <number>, <identifier> and <string>
respectively.

<string> is a special case because it can have whitespace inside it,
but the entire "foo bar" thing is a single "object". So, now, let's
take a closer look at what's going on here.

The rule (in part) <comment> <comment> <comment> worked before the
change, because it was *really* <comment> \s* <comment> \s* <comment>.
But, when we changed this from <comment> <comment> to <comment>+,
Perl 6 assumes we know what we're doing and stops inserting whitespace.

So it now "expands" to <comment> <comment> <comment> again, without
the \s* in between. There are other ways to turn the whitespace
behavior back on, but we'll save those for later in the grammar.
For now, it's easy enough to insert the whitespace at the end of
the original token.