Moving a little faster now...
Let's tackle a little bigger bite now. I'm going to take a guess
that 'import modulename.' is picat's way to import a module. Seems
a bit risky to me, but I'm guessing that's what's happening here.

So, let's encapsulate this bit. First, we can call this three terms
in a row. 'import' is what we'd call a "keyword." 'modulename' is...
well, a module name, and '.' is the end of.. well, what I guess we
can call a statement, for lack of a better word.

I'm going to guess it's the equivalent of 'use Module::Name;' in Perl,
and call the whole thing "import-statement". So, according to the
rules we came up with in our last commit, 'import' and 'modulename'
can probably be separated by any amount of whitespace we want,

So, these things should be part of a rule and not a token. We could
just jump in and write:

rule import-statement
  {
  'import' \w+ '.'
  }

and capture 'import module_name .' ad nauseam, but that wouldn't be
as fun as what's coming up. 'modulename' here should really be
given its own name here, so let's create a ... not a rule, but a
token for that.

This is also handy because we could (though it's highly unlikely)
find out that module names can be arbitrary strings, and we don't
want to have to come back and change our code.

So, if we write:

token module-name
  {
  \w+
  }

and use that in place of our \w+ above. That also gives us a name
for what "arguments" 'import' takes, and reads clearer.