Adding a function name
There are a bunch of things I could have done here, but of course
I chose this. There's not much except for the real meat of things,
the function bodies and a few random function calls. I'll want
eventually to create something like a function rule, and it'll
likely look something like this:

rule function-definition
  {
  <function-name> <function-arguments>? <function-body> '.'
  }

and the first term of this happens to be <function-name>. It's
simple, but it'll work, and is a simple way to start in on
our first "real" rule.

Now, this *does* actually change things slightly, And in recognition
of this I'm going to add a regression test as a reminder that
we've potentially broken our grammar. It shouldn't cause much of
a problem, but admittedly these things do tend to build up, and
it's a good idea to keep an eye on them, or at least acknowledge
the fact that there may be a problem here.