Adding a "regression" as a note more than anything...
I've got a whole corpus of test files to look through, I'll check
that to see if any of the function names has spacing between the
last character of the name and the opening parentheses. I basically
forgot about that possibility, but part of this git log is
recording both the mistakes and the mainline of changes.