Clean up some of the terms, separate out '.' for later use
Basically getting my head back into this. I'll note in passing that the
first time I ran this test it *might* have "infinilooped", shorthand
for "got stuck." I removed the lib/.precomp directory just in case
and it ran just fine afterwards.

I think the next thing to do won't advance the grammar per se, but it
will add a litle more structure, and it'll create a term that's
going to see a lot more use.