Create the basis of a "statement".
It might be about time to talk about naming things. Language design
has a long and storied history, and you'll see that some parts have
somewhat confusing names.

Like "statement" here. I've combined <import-statement> and <comment>
here into one "statement" with the idea of being able later on to
add things like "assignment-statement" to it later on. But one thing
that you might be thinking is that "But...but... a statement *states"
something, a comment really doesn't have anything to say."

This is true, but comments behave like <import-statement>
syntactically. They can be on their own line like an assignment,
and they can kind of "hang out" after the import-statement,

The comma is a statement separator, and if you look at lines like:

'initial_state([2,4,1,7,5,3,8,6])' '.' <statement>+

so is a period. So it can take the place of a statement, like this:

   write(L), nl,

which also has two statements in a row. It looks like a duck
(it can take the place of a statement), it quacks like a duck
so I'll call it a duck. True, you can't stick another statement
after it and expect it to compile and execute, but that's just
kind of a quirk of the comment tag - besides, we still allow

  write(L), /* foo */ nl,

to compile, although not many people use this because it's hard to
read.