Break apart statements at separators
These would be the ',' and '.'. All I've really done here so far is
(with a few exceptions) separated out the statements. If you've got
a good eye, you might see that I seemed to miss a few things. Those
are intentional, actually. (unless I missed one, in which case...
wait until the next commit.

Take a quick look at:

  'foreach(I in 1..N)
     Root = sqrt(I)' ','
     'writeln([I, cond(Root == 1.0*round(Root), open, closed)])
  end' ','

You'll see that I broke out the ',' but I didn't break out the
'foreach(I in 1..N)' bit above it. This is mostly to remind me that
I really don't think "foreach(...)" is a statement in itself, but
in fact it can better be thought of as a "header" to a block.

In Perl 6, we have constructs like:

while {
  #`{ Code here.. }
}

When I was talking about statements earlier, you might remember that
I said I considered things a statement if they could be run stand-alone.
Well, 'foreach(I in 1..N)' has a problem. You can't really run that
without knowing what code you're foreach'ing over.

So the full 'foreach..end' thing is kind of a statement in that the
entire thing is what gets compiled, but it's got statements *inside*.
How we usually do this when doing compiler stuff is we'd treat it
like:

rule foreach-statement
  {
  'foreach' '(...)' <statement>+ 'end' '.'
  }

If the language were more block-structured, then you could even do
something fancy like:

rule foreach-statement
  {
  'foreach' '(...)' <block>
  }

rule block
  {
  '{' <statement>* '}'
  }

with the advantage that you can reuse the '{'..'}' bit for other
constructs like:

rule with-statement
  {
  'with' <variable> <block>
  }

In passing, I'll note that I've done enough work with this that I can
read what's up there at sight and be able to tell what it should be able
to parse.

I just walk left-to-right and mentally translate the terms, so I'd
read that as "with $foo { ... }" without really having to reference
other lines. Variables are a compact way of "understanding" "$foo",
"@bar", "$bar[32][$ding]" in other languages.

I also have a list of corner cases there, like $bar[32][$dig] that I
mentioned, so I can run over the list to see if there's a combination
of variables and the 'with' construct that doesn't work together.

A better explanation of that might be:

rule modulus
  {
  <variable> '%' <constant>
  }

'$x % 32' is valid, so is '@foo[3] % -7'... but '$x % "FOO"' wouldn't
make sense, because while 32 and -7 are constants, so are things
like strings. That wouldn't work together, so I might want to try
something like this:

rule modulus
  {
  <variable> '%' [ <number> | <variable> ]
  }

But it's past my cutoff time, and I'm going to turn into a pumpkin
if I don't commit and get to bed.