Breakin' up isn't hard to do
Break up all of the statement blocks, now that I've made the point
that there will be some things in there that won't be statements.
I'm sure that the language has more than just "foreach" loops.

So, at least the 'end' terms won't be statements on their own. We'll
keep an eye out for those, and maybe even do those on the next pass.

Nothing else really changed, although I should point out again that
if 'nl .' is illegal in the actual language, that we have changed
the actual spec a little bit. Again, I think it'll be simple to remove
the space if I need to, probably by changing from a rule that makes
the whitespace optional to a token that makes the whitespace illegal.