Cheating with functions...
This is actually not what I'd do while designing this grammar. I'd
actually jump ahead to what I know is coming up, but this is meant
to be both a git log and a tutorial of sorts.

I'm also seeing a potential issue, because what I want to do is
be able to replace 'go' with "zero or more statements" eventually,
but that means that 'function_name => import x.' will be legal,
because 'import x' is a legitimate statement right now.

That doesn't feel right, so I've just consulted the official grammar.
Yes, there is one, in the user manual. I haven't done that before,
but tonight felt like the right time.

It says that there's an overarching 'program' construct that has an
optional (that's what the [] around it in the grammar means - it's
a BNF grammar, and that's how they call options.) module declaration
then an optional import statement.

So next thing will be to fix that mistake, and then we can get on
with the proper buildup.