Fix up imports
This comes in two pieces. The first and most obvious change is that
'import foo' is no longer an official statement. If it were, that would
mean that "import foo" could occur anywhere in the file, and I now know
that's not legal. So out it goes, and I have to restore a bit of the
earlier grammar.

I've also made a slightly more subtle change, and reverted what were
<statement>s above the <import-declaration> to their original <comment>
name. This may be a sign that <comment>s really won't be part of
the <statement> rule in the long term, but I'll reserve judgement
on that.

Also, the official grammar says that a 'module' declaration exists
above the 'import' declaration, but I'm not testing that just yet.

I could do that easily, and add another test... I think I might just
do that as the next step.