Another woopsie found and fixed...
Tiny little change - I forgot that I'm testing this grammar as the
*entire* file, so I restored the ^ at the top of the blocks.

In the next step I think I'm going to do a little more
infrastructure work and build out a proper <program> declaration
block. This will also move the bulk of what we're working on
here down one level.

It'll be a big change, but what's going to happen is that TOP will
no longer contain the entire file worth of matching, instead it'll
look roughly like this:

rule TOP
  {
  ^ <program> $
  }

rule program
  {
  <comment>* <import-declaration>?

    [
    ... first file's worth...
    ||
    --- second file's worth...
    ]
  }

Yes, this does look much like the official grammar, and that's
intentional. Naming is hard, and if I have an established guide
I'll follow that. And what's a better guide than the official
reference?

I've not followed it until this point simply because I can, and
I can show that you can do a lot on your own without a guide.