Refactored so that I can keep import statements separate.
Now that I know that 'import' and 'module' can only occur at the
start of the module, I've refactored things so that 'import foo.'
is in the main TOP block, so it won't participate in the body
of the program.

This is also why I renamed 'import-statement' to 'import-declaration'.
I'm more used to Perls 5 and 6 where 'use' statements can occur most
anywhere in programs. This distinction is important in Perl 6 where
your import statements can affect the syntax after it's been included,
but in picat apparently it's really not.

I've also named the "child" rule 'program-block' because I might
get around to creating an official <program> block at some point.
In BNF "program" (or by convention the last rule) is the root of
the grammar. In Perl 6, TOP takes that role on.

It's also easier to do this in order so I don't have to flip to the
bottom of the file to see what the root declaration is.

Next time we'll get back to the (sort of abandoned)
<function-declaration> thing, and actually make it work.