Bare function calls are statements...
Terms like 'go' and 'nl' are function calls, and much Perl 5/6,
they don't need parentheses. So let's go through and replace all of
the function calls with an actual <statement> rule.

The idea is that in a few revisions, most of the function bodies
will be <statement>+. We'll compress those to <function-declaration>
which has a little glitch in it at the moment.