Woops, forgot to *actually* factor out <import-declaration>
That's one of the nice things about this strategy. It's resilient
in the sense that I can collapse a string of terms into a single
rule and forget to propagate that change throughout the code.

In fact that particular change probably has been unnoticed even by
me for a little while, the main reason being that it's quite a ways
down in the file.

It would scream bloody murder if any of those terms wasn't optional,
but at least you'd know what change you made. This is one reason
why I'm committing one change at a time, that way I only have to
test one thing at a time.

It's not perfect, and it does unfortunately leave quite a bit of
clutter, but at least you can look at small changes instead of
going for it and then wondering what out of all the terms and rules
which one actually broke things.