Layers upon layers...
I've added a layer around <function-name> in preparation for
something coming up. I'm seeing something coming up here quickly,
but you might not see this. You might be able to see what's going
to happen if you analogize picat to Perl.

In Perl, 'exit;' and 'print("hello");' are both function calls,
it just happens that you don't pass arguments in one. So, function
calls might not be *just* a function name.

So I'll start out by creating a <function-call> rule because I know
that <function-name> isn't the only way to call a function. I'm going
to use this next, but I thought it was easier to take this in
little steps.

We could do this in a bunch of ways, but this is pretty common.
There's a further reason that I probably shouldn't get into, but
I'm going to create an argument list to capture "(100)", "(foo,I[])"
like things.

Picat doesn't have this separation, but Perl 6, Pascal, C, and
Java among others has a notion of both an argument list like

foo( 100 );

and a formal parameter list like

sub foo( Int $a, Str $b );

So, Picat doesn't seem to distinguish between argument lists and
parameter lists, but I may be wrong. So I want to keep parameter
lists and argument lists in different rules, just in case. Because
this can cause problems later on, I'd just as soon avoid it by
keeping the categories separate.