Look at this one carefully...
Being able to call 'foo(123)' just removed a bunch of statements, and
one you might not have expected. Look down for 'L =' and you'll see
that I've created a <function-call> inside what's eventually going to
be a <statement>.

  'L = new_list(10),'

is a statement, to be sure. It just happens to have a <function-call>
in the middle of it. We'll pick up the assignment portion of this
later on, but for now we'll leave this as a reminder that assignments
are statements too.

We'll generalize this even more next commit.