Is this the right commit for an argument?
I've told you once...

Anyway... you'll see with this little change that a *bunch* of
our original text just went away and was replaced with <statement>
and <function-call>. If you'll take a close look at the diff,
you'll see that I use both of these.

Basically, I use <statement> if I can replace from the '=>' or ','
to the next ',' or '.', otherwise I use <function-call>. In Perl
terms, this is a statement:

function(12);

But function calls can be found in statements too...

$x = function(12);

I'd replace *this* instance with <function-call>. Basically, I know
that this is going to be subsumed into <statement>, and I might
actually do this next. But you'll see that lots of text just got
replaced with <statement>, so we've just made a big jump ahead.