Another refactor to clean things up.
One common guiding rule when creating grammars is
"Don't repeat yourself." We were kind of doing that before, in this
segment:

  {
  | <function-name> '(' \w+ ')'
  | <function-name> '(' \d+ ')'
  }

Everything but \w+ and \d+ is repeated in this, and the \w+ and \d+
(like 'Doors' and '10') terms are really arguments. So name them
arguments.

And in this case create a token, not a rule. Because these things
should just be one term. If it were a *parameter* list, this would
look much different, maybe something like this:

rule parameter
  {
  | <function-name> '(' \w+ \w+ ')'
  | <function-name> '(' \w+ \d+ ')'
  }

and this would be able to match 'foo( Int ab )' and 'foo( Const 32 )'
parameter lists. That's why '('..')' gets treated specially. Next, we'll
stop repeating <statement>s.