Collapse another function-definition.
I'm introducing a new bit of grammar here in order to pull off
this bit of wizardry. It's a simple change, but needs some
explanation.

  <statement>+ %% ','

This just acts like <statement>+ in that it lets us string a bunch
of <statement>s together, but the "%% ','" means that we string
this bunch of <statement>s together, but stick ',' in between each.

If I just said <statement>+, then I'd only be able to string
"nl go do_form(I)" together. Asding "%% ','" allows me to stick commas
between each of those terms, so "nl, go, do_form(I)" works. Whitespace
around the commas are flexible.

We haven't collapsed all of the function bodies yet, but we'll slowly
do that as we go through these things.