Fight for your right...hand expression.
I know, that was very much forced. But it's late, and about time to
pack it in. But you'll see with this change that I've snuck something
in under the radar.

Look carefully at the new <statement> section:

  rule assignment-statement
    {
    <variable-name> '=' <function-call>
    }

  rule statement
    {
    | <assignment-statement>
    | <function-call>
    | <comment>
    }

You'll see that <statement> "calls" function-call, *and* the
<assignment-statement> is calling it. If you know what's going on in
terms of yacc-style grammars, then look at it this way:

  rule assignment-statement
    {
    | <variable-name> '=' <function-call>
    | <function-call>
    | <comment>
    }

and you'll see that I've got the same term twice on the right-hand
side. This is perfectly okay, but if I were to write something like
this:

    | <function-call> '=' <variable-name>
    | <function-call>
    | <comment>

you'd scream "left-recursion!" and be right. (ha!) But that's where
one of the differences between traditional parsing algorithms and
Perl 6 comes into play. '|' forces the grammar to look for the longest
match, so it would try to match 'x(1) = foo' before 'x(1)'. So it'd
still work, but seem counterintuitive.

I could go on and on about this change, but that might end up being
an entire tutorial course in itself. But enjoy the change, and watch
how we're whittling away at the statement blocks.