Fix a few more cases where <statement>s could be added...
Nothing big, I just missed a few cases where 'Len)' was hiding the
fact that there was an assignment expression. And come to think of it,
remember, naming is one of the hardest things, in computer science,
 just noticed that 'foo(x = y)' is valid, and more to the point, it's
different than 'foo(x == y)' so comparison and assignment are different.

They aren't in some languages, see SQL for counterexamples. I think I
should do a quick rename next...