As promised, just a quick cleanup.
Mentioned earlier that 'foo(x = y)' is legal, as is 'foo( x == y )',
so like Perl and most other languages, you can do assignents inside
expressions. So <assignment-statement> really should be called
<assignment-expression>.

Which gives us the seed for the ultimate <expression> rule, which is
where the *real* meat of the application is focused.