Propagate assignment-expression across the dataset.
This is hopefully a trivial change if you've been following along,
the RHS of the assignment-expression isn't quite yet fleshed out,
because arrays are an important type to have. Next change will
probably be comsmetic, I'll reformat the lines so they look nicer
and we can see how many repeating <statement> ',' pattterns we
really have in the code.