Record-scratch moment here.
I'm going along from the bottom to the top replacing [M1,...] constructs
with my newly-minted <array> rule, when I see something that stops
me im my tracks... sort of.

This is the line:

   'time(plan3(Init,L,Cost,[]))' ','

Note this is ... a bit complex.

But when I break it apart into what it should be, at least in my head,
I get this parse:

   'time(' 'plan3(' 'Init' ',' 'L' ',' 'Cost' ','' []' ')' ')' ','

I'm not going to worry about most of it right now, but I should talk
for a moment about the '[]' portion of this. I initially defined
(as you can see) <array> as:

  rule array
    {
    '['
    <argument>+ %% ','
    ']'
    }

The "%% ','" construct just tells us <argument>s are separated by
required commas and optional whitespace, you might remember this from
the original <argument-list>.

I just happened to notice as I was going through [..] updates that this
was an empty array ref, appearing in an argument list. For a moment
I thought it was a nested array, which may actually be legal in picat
(and probably is), but it's not