Ouch...
As promised, this is truly a blow-by-blow account of how this
translation goes down. You'll notice if you take a quick look at the
diff that I've added <function-call> as a possible argument type
at the top of the list of  <argument>s.

And it works fine for <function-call <function-call> > things, but...
not <function-call <statement> > things. Which is actually perfectly
understandable, but here's the clinker.

I could easily ignore this and go on, but I think it's about time
for a confession/refactoring. Because at least in the world of grammar
design, a statement *really* isn't a statement until you're at the end
of the statement block.

For instance, in Perl 5/6, you can have a bunch of things that would
be statements in their own right, like so...

    print 1 + 2, 3 + 4, 5 + 6 ;

the '1 + 2', '3 + 4', and '5 + 6' specifically. But the entire thing
really isn't a statement _per_se_ until we hit the trailing ';'.
And I've been calling things statements that really aren't throughout
this discussion. This has partially been so that I don't have to
explain too much at any given point, and partially sheer laziness.

Now it's time to fix things. I'm also doing this, at least from an
exposition standpoint, to show you that you can keep design mistakes
in the grammar for quite a while until the bite you in the ass.

So if you'll excuse me for a few minutes, I'm going to flip back
to the other tmux window to figure out how best to do this in just
one or two commits.