Restate what a "statement" is.
This is actually a bigger deal than it might appear at first.

I'm going to follow the lead of ... well, most other languages out in
the wild, and "arbitrarily" state that a "statement" is an expression
followed by a comma. And in fact, to clear that up a bit, I'm going to
declare a <comma> type and use that everywhere a comma appears.

A statement can also appear at the end of a function block, and if so
it's followed by nothing. This is primarily because the '.' looks like
it's tied to the function definition. Every place a '.' appears, before
it there's a list of comma-separated statements, then a '=>'.

So I'll make these easier to read, first off.