Replace '.'...
And a behind-the-scenes whoopsie. I accidentally flipped from vim
over to git, did an up-arrow and <return>, which unfortunately
ran 'git reset --hard' rather than '6 t/01-basic.t'. But again, I
had the original file that I'd just edited in the other window.
So I could just save out the file with my edits rather than redo my
old work.

But anyway, I've added <period> to the grammar to make things easier
to read, because in most fonts, '.' and ',' are hard to tell apart.
Something I should have done a while ago, but this also makes the point
that a statement should only appear after a comma or period.

Now I'll do the same thing with ',' => <comma>, and we'll see things
just a little clearer.