Comma now...
If you'll look closely, the file finally follows our rule that
a statement can only be followed by a <comma> or <period>, and since
the <period> is part of the <function-definition> it really doesn't
count. So in the end a statement is either followed by <comma> or
nothing.

There is a subtlety hiding here. If you'll look at our code, this
section looked a bit different before:

<function-name> '=>'
   <comment>
   <statement> <comma>

It used to read:

<function-name> '=>'
   <statement>
   <statement> ','

But because of our rules, this would be a <statement> followed by
another <statement>, and a syntax error. This rule *may* ring some
proverbial bells. If you've spent more than 10 minutes programming
Perl 6, this error may look familiar:

"Two terms in a row..."

This almost the same error, but in a different language. In
Perl, it comes when you try to put two terms together without a
spacing semicolon in the way:

exit()
while(...)

So the "term" 'exit()' and 'while(...)' are together without the
required spacer ';'. In the case here of the picat grammar, it's
<statement> <statement>. But what's different here is that the first
<statement> really isn't a <statement>, it's a <comment>. And the
rules here are so precise that even without looking at the code, I
could tell that the first <statement> needs to change to a <comment>.

So, now you'll see that (eventually - there are a few hangers-on here)
every <function-definition> follows the same pattern of

  <function-name> <arguments>? ( '=>' or '?=>' )
    <statement>+ %% <comma>
  <period>

so eventually that'll be how pretty much the entire file looks, just
rows of <function-definition> blocks.`