[Broken] Let's make sure all of our <comment>s are out...
by commenting out the <comment> line in <statement>s.... Woops.

Looks like we've forgotten a few. But I don't want to compare every
line with the original files, especially with all of the changes
we've been doing.

So let's not, instead let's use Grammar::Debugger... Aha, that tells
me that the first <statement>+ block isn't parsing correctly, so it
must be <comment>+ instead. That's fixed the first test...

For the next test I tried the same idea but ran out of patience quite
a ways in, and just searched for 'index(-)' in the original file. Found
it... two blocks that are just more comment blocks.

Grammar utilities for debugging are a little problematic, but it's
not really the language's fault. They're compiled now, so even external
authors don't get much to work on, just a bunch of compiled methods.

I've got a few ideas on how to fix this situation, but they're very
long-term. This commit message is just to reassure you that you're not
alone when things like this happen, and you can get through it.