Finally refactor out the <expression>...
It's now finally just a change of name. Change what was a <statement>
to an <expression>, just what it truly is in the actual grammar. This
also makes it a little clearer to see what the next change is going
to be.

I'll add two cases that don't yet occur in the grammar just to be sure
we've covered the possibility.

Suppose we have something like this:

my_func =>
  % flying comment
  /* embedded comment */ nl, % flying
  /* embedded comment */ nl,
  nl, % flying
  nl,
  nl
.

So this would collapse right now to something like this:

<function-name> '=>'
  <comment>
  <comment> <expression> <comma> <comment>
  <comment> <expression> <comma>
            <expression> <comma> <comment>
            <expression> <comma>
            <expression>
<period>

I've lined things up around the <comma> so we can see what's going on.
It looks like there's a lot of possible cases to cover, so let's
try to break things down.

rule thingie
  {
  | <comment>
  | <comment> <expression> <comma> <comment>
  | <comment> <expression> <comma>
  |           <expression> <comma> <comment>
  |           <expression> <comma>
  |           <expression>
  }

This would definitely cover it, so let's try it out.