Treat <array>s as arguments.
This trims things down quite a bit as well, nothing really interesting
although it's of note that we're widening the amount of things that can
be included in argument lists.

You'll also note I'm not including what look like list comprehensions
in the 'foreach' terms. Next on my list is comparisons, so we can get
an <if-expression> together.