Fix <function-definition>, apply it.
Nothing special to note here, <function-definition> didn't account for
the fact that 'foo(Bar,1) => ' style calls existed.

You might be wondering why I didn't jst change the definition to accept
the weird-looking '?=>' style marker and get a big chunk of the definitions
straightened out.

It's precisely because it "looks weird", and I know enough Prolog to
know that that's probably a fact, not a function. That may not be the
correct name for it, but it should be easy to fix later on.

Yeah, I'm piling potential problem on top of potential problem, but that's
part of develpment too. I'm keeping these things in mind, and if an
assunption I made breaks something down the line I'll eat the proper
amount of crow and fix it.