Here's some interesting breakage...
I'd forgotten about these lines, and just changed one of them
because I assumed that 'foo(' <expression> ')...' was a legitimate
<function-call> type. Turns out that it's not.

The interesting question is how to properly fix it. Here's the section
of the grammar that needs reworking:

  rule argument
    {
    | <function-call>
    | <array>
    | \w+
    | \d+
    }
  # ...

  rule function-call
    {
    | <function-name> '(' <argument>+ %% <comma> ')'
    | <function-name>
    }

  rule expression
    {
    | <assignment-expression>
    | <function-call>
    }

Now, 'writeln(' <expression> ')' is what's breaking in this case. But
why, and where should the fix be applied? I could patch things by
adding <expression> to <function-call> like this:

  rule function-call
    {
    | <function-name> '(' <expression> ')'
    | <function-name> '(' <argument>+ %% <comma> ')'
    | <function-name>
    }

and it would work, but it wouldn't be quite right. That would only allow
'writeln(foo=1)' or 'writeln(foo,1,bar...)'. It's entirely reasonable
to think that 'writeln(foo=1,bar,...)' is legal as well. Or even better,
our <expression> could instead be 'foo(bar==1)', and in that case, allowing
multiple Boolean expressions in a single function call is a no-brainer.

So, if I wanted to allow 'bar==1' in an argument list... Well, there
really is one sensible place to allow this, and that's as an argument.

  rule argument
    {
    | <function-call>
    | <array>
    | \w+
    | \d+
    }

becones

  rule argument
    {
    | <function-call>
    | <expression>
    | <array>
    | \w+
    | \d+
    }