Collapse the 'writeln(' <expression> ')' bits
And you'll notice we've collapsed another <function-definition>.
If you're following along as well, you'll notice that I put the
<expression> at the top of the list. We're going to refactor that
area pretty soon, but for the moment we'll go this route.

The reason this didn't work is a little obscure, but you do need to
know what's going on here. The parser is starting with:

         |
         V
  'write(len=Len)'

It's inside a <function-call> match, looking for an <argument> term.
So it's got a bunch of alternatives, the first one just happens to be
what we just left, <function-call>. Now, that does happen to match this
bit of the string

         <function-call '(' <argument <function-call <function-name ...
            |
            V
  'write(len=Len)'

But it gets to the equal sign, and the parser fails. Whereas if we put
the <expression> first, which is probably what it should be in the first
place, it immediately matches expression -> assignment-expression and
life goes on smoothly.

  rule argument
    {
    | <function-call>
    | <expression>
    | <array>
    | \w+
    | \d+
    }

  rule function-call
    {
    | <function-name> '(' <argument>+ %% <comma> ')'
    | <function-name>
    }

  rule assignment-expression
    {
    <variable-name> '=' <argument>
    }