Whew, *big* refactoring.
But it got us some important stuff.

The big thing for me is that the <expression> term, even though it
just looks like the <argument> term rearranged, now has the right
contents.

In general, the <expression> term is the hardest to write in any
language, because it has to do the heavy lifting. Almost every
statement contains at least one expression, and every expression is
where something gets *done*.

So it's nice to see the <expression> term lining up so neatly.
It's going to be very much recursive, but the important thing is that
the left-hand side of the expression should be as unique as we can
make it. You'll see more as we go along.

The rest of the changes play into that. Function calls now take a
list of expressions rather than generic arguments, arrays do the
same. It's really just a renaming of the term, but it should clear
up quite a few things.