Whuh-oh.... Better be careful with this one.
It'll loop forever if you let it.... workin' 9 to 5

This... is a problem. Remember how I mentioned earlier that most
expression grammars kept the left-hand side as unique as possible? Well,
you've just found out the reason.

First of all, why does it do this? Well, let's play along here.

Here's the offending section of the grammar:

  rule comparison-expression
    {
    | <expression> '<=' <expression>
    }

  rule expression
    {
    | <assignment-expression>
    | <comparison-expression>
    | <function-call>
    | <array>
    | <variable-name>
    | \d+
    }

and let's assume that we're working on this phrase:

 expression
 |
 V
'a <= 10'

With the "cursor" at the V marker, a little ASCII graphics.

The parser checks each alternative in turn. <assignment-expression>
matches until this point>

 expression
 assignment-expression
   |
   V
'a <= 10'

When the expected '=' fails to be there, so the parser backtracks and
tries the next alternative...

 expression
 comparison-expression
 |
 V
'a <= 10'

And the loop starts. Because when the <comparison-operator> term
runs, what's the first term?

 expression
 comparison-expression
 expression
 |
 V
'a <= 10'

And the same thing happens over and over until the program runs out of
stack space.

So....how do we get out of this situation?

Well... let's see here. We can stop the recursion easily because we
actually did more work than we really needed to. We're really just trying
to match a variable-name, not an expression. So let's start off fixing that.

Change

  rule comparison-expression
    {
    <expression> '<=' <expression>
    }

to

  rule comparison-expression
    {
    <variable-name> '<=' <expression>
    }