This fixes one part of the expression...
So 'a <= 5' and 'a[I] <= E' will work with this arrangement, because
a <comparison-expression> is <variable> '<=' <expression>. But so
much more won't work... pretty much anything on the left-hand side.

In C, a common mistake is typoing '=' for '==', so quite a few of us
have gotten into the habit of writing '3 == a' instead of 'a == 3'
because if I forget an '=' sign, this will compile into '3 = a'
and generate a syntax error trying to assign a value to a number.

So, we probably should allow this in our grammar. Let's do some
more refactoring to allow both 'a <= 5' and '5 <= a' as legal
expressions.