Now, how we get out of this pickle...
Remember that we need to keep the left-hand side reasonably unique.
It doesn't have to be *completely* unique, just that it'll stop the
recursion.

Let's now go through what we've written on the same two statements
'a = 10' and 'a <= 10'.

  rule assignment-expression
    {
    | <variable-name> '=' <expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <variable-name> '<=' <expression>
    }

  rule expression
    {
    | <assignment-expression>
    | <function-call>
    | <array>
    | <variable-name>
    | \d+
    }

 expression
 |
 V
'a              =   10'

So this matches the very first alternative...

 expression
 variable-name  '=' expression
 |              |   \d+ <- the last alternative in <expression>
 |              |   |
 V              V   V
'a              =   10'

So we handle the first case. Now, the next case is where the
recursion started the last time. So, how do we avoid it this time?

Like the first time, we walk up to the '<'...

 expression
 |
 V
'a <= 10'

 expression
 variable-name
 |
 V
'a <= 10'

   expression
   '='         <-- Beep! beep! failed match! Try next alternative!
   |
   V
'a <= 10'

 expression
 comparison-expression
 variable-name
 |
 V
'a <= 10'

   expression
   comparison-expression
   '<='
   |
   V
'a <= 10'

So now we've gotten rid of the infinite loop, each of our cases has
at least one token that's unique to it somewhere, and they don't
involve recursing from one term back into the original term, causing
another infinite loop.

This also gives us a way to separate assignment and comparison expressions
which is important when we write the compiler, as we *have* to know which
type of expression comes first. This is important because if we don't put
them in the right order, 'a = 10 <= 5' is clearly different than
'a <= 10 = 5' and we don't want to be rewriting the user's code
differently.